+++ Src/AutoSpectre.SourceGeneration.Tests/ChoicesTests.cs
@@ -4,6 +4,63 @@ using Xunit.Abstractions;
 
 namespace AutoSpectre.SourceGeneration.Tests;
 
+public class ConverterTests : AutoSpectreGeneratorTestsBase
+{
+    public ConverterTests(ITestOutputHelper helper) : base(helper)
+    {
+    }
+
+    [Theory]
+    [InlineData("public static string PropertyConverter(DateTime input) => string.Empty;",".UseConverter(TestForm.PropertyConverter)")]
+    [InlineData("public string PropertyConverter(DateTime input) => string.Empty;",".UseConverter(destination.PropertyConverter)")]
+    public void ValidConverterSinglePromptGeneratesExpectedResult(string converter, string expected)
+    {
+        GetOutput($$"""
+                    using AutoSpectre;
+                    using System.Collections.Generic;
+                    using System;
+
+                    namespace Test;
+
+                    [AutoSpectreForm]
+                    public class TestForm
+                    {
+                         [SelectPrompt()]
+                         public DateTime Property {get;set;}
+                         
+                         public List<DateTime> PropertySource {get;set;}
+                         
+                         {{converter}}
+                    }
+                    """).OutputShouldContain(expected);
+    }
+    
+    [Theory]
+    [InlineData("public static string PropertyConverter(DateTime input) => string.Empty;",".UseConverter(TestForm.PropertyConverter)")]
+    [InlineData("public string PropertyConverter(DateTime input) => string.Empty;",".UseConverter(destination.PropertyConverter)")]
+    public void ValidConverterMultiSelectPromptGeneratesExpectedResult(string converter, string expected)
+    {
+        GetOutput($$"""
+                    using AutoSpectre;
+                    using System.Collections.Generic;
+                    using System;
+
+                    namespace Test;
+
+                    [AutoSpectreForm]
+                    public class TestForm
+                    {
+                         [SelectPrompt(Converter = nameof(PropertyConverter))]
+                         public List<DateTime> Properties {get;set;}
+                         
+                         public List<DateTime> PropertiesSource {get;set;}
+                         
+                         {{converter}}
+                    }
+                    """).OutputShouldContain(expected);
+    }
+}
+
 public class ChoicesTests : AutoSpectreGeneratorTestsBase
 {
     public ChoicesTests(ITestOutputHelper helper) : base(helper)
@@ -131,7 +188,7 @@ public class ChoicesTests : AutoSpectreGeneratorTestsBase
                          
                          {{invalidSource}}
                     }
-                    """).ShouldHaveSourceGeneratorDiagnostic(DiagnosticIds.Id0023_ChoiceCandidate_NotValid);
+                    """).ShouldHaveSourceGeneratorDiagnosticOnlyOnce(DiagnosticIds.Id0023_ChoiceCandidate_NotValid);
     }
     
     [Theory]
@@ -155,7 +212,7 @@ public class ChoicesTests : AutoSpectreGeneratorTestsBase
                          
                          {{invalidSource}}
                     }
-                    """).ShouldHaveSourceGeneratorDiagnostic(DiagnosticIds.Id0023_ChoiceCandidate_NotValid);
+                    """).ShouldHaveSourceGeneratorDiagnosticOnlyOnce(DiagnosticIds.Id0023_ChoiceCandidate_NotValid);
     }
     
     [Theory]
@@ -177,7 +234,7 @@ public class ChoicesTests : AutoSpectreGeneratorTestsBase
                          
                          {{invalidSource}}
                     }
-                    """).ShouldHaveSourceGeneratorDiagnostic(DiagnosticIds.Id0024_ChoiceCandidate_NotFound);
+                    """).ShouldHaveSourceGeneratorDiagnosticOnlyOnce(DiagnosticIds.Id0024_ChoiceCandidate_NotFound);
     }
     
 }